Fix issue causing Medium open publication action v2 script to not work.
Because this action uses includeHeaders, the string passed to the script isn't just a URL, it's a URL+newline+JSON of headers. That meant that fetching the URL wasn't working, which made this never work. The v1 JS executor is more lenient with things like this, so the issue didn't exist there.